+++ webui/module/Job/src/Job/Controller/JobController.php
@@ -5,7 +5,7 @@
  * bareos-webui - Bareos Web-Frontend
  *
  * @link      https://github.com/bareos/bareos for the canonical source repository
- * @copyright Copyright (C) 2013-2022 Bareos GmbH & Co. KG (http://www.bareos.org/)
+ * @copyright Copyright (c) 2013-2022 Bareos GmbH & Co. KG (http://www.bareos.org/)
  * @license   GNU Affero General Public License (http://www.gnu.org/licenses/)
  *
  * This program is free software: you can redistribute it and/or modify
@@ -569,8 +569,6 @@ class JobController extends AbstractActionController
 
     $this->bsock = $this->getServiceLocator()->get('director');
 
-    // TODO
-
     return new ViewModel();
   }
 
@@ -598,6 +596,7 @@ class JobController extends AbstractActionController
     $data = $this->params()->fromQuery('data');
     $jobid = $this->params()->fromQuery('jobid');
     $jobname = $this->params()->fromQuery('jobname');
+    $jobs = $this->params()->fromQuery('jobs');
     $status = $this->params()->fromQuery('status');
     $period = $this->params()->fromQuery('period');
     $client = $this->params()->fromQuery('client');
@@ -636,6 +635,24 @@ class JobController extends AbstractActionController
         echo $e->getMessage();
       }
     }
+    elseif($data == "all-job-resources") {
+      try {
+        $jobs_B = $this->getJobModel()->getJobsByType($this->bsock, 'B'); // Backup Job
+        $jobs_D = $this->getJobModel()->getJobsByType($this->bsock, 'D'); // Admin Job
+        $jobs_A = $this->getJobModel()->getJobsByType($this->bsock, 'A'); // Archive Job
+        $jobs_c = $this->getJobModel()->getJobsByType($this->bsock, 'c'); // Copy Job
+        $jobs_g = $this->getJobModel()->getJobsByType($this->bsock, 'g'); // Migration Job
+        $jobs_O = $this->getJobModel()->getJobsByType($this->bsock, 'O'); // Always Incremental Consolidate Job
+        $jobs_V = $this->getJobModel()->getJobsByType($this->bsock, 'V'); // Verify Job
+        $jobs_R = $this->getJobModel()->getJobsByType($this->bsock, 'R'); // Restore Job
+        $result = array_merge(
+          $jobs_B,$jobs_D,$jobs_A,$jobs_c,$jobs_g,$jobs_O,$jobs_V,$jobs_R
+        );
+      }
+      catch(Exception $e) {
+        echo $e->getMessages();
+      }
+    }
     elseif($data == "details") {
       try {
         $result = $this->getJobModel()->getJob($this->bsock, $jobid);
@@ -660,13 +677,13 @@ class JobController extends AbstractActionController
         echo $e->getMessage();
       }
     }
-    elseif($data == "timeline") {
+    elseif($data="job-timeline") {
       try {
         $result = [];
-        $c = explode(",", $clients);
+        $j = explode(",", $jobs);
 
-        foreach($c as $client) {
-          $result = array_merge($result, $this->getJobModel()->getClientJobsForPeriod($this->bsock, $client, $period));
+        foreach($j as $jobname) {
+          $result = array_merge($result, $this->getJobModel()->getJobsForPeriodByJobname($this->bsock, $jobname, $period));
         }
 
         $jobs = array();
@@ -736,7 +753,7 @@ class JobController extends AbstractActionController
               $endtime += 1000;
             }
 
-            $item = '{"x":"'.$job['client'].'","y":["'.$starttime.'","'.$endtime.'"],"fillColor":"'.$fillcolor.'","name":"'.$job['name'].'","jobid":"'.$job['jobid'].'","starttime":"'.$job['starttime'].'","endtime":"'.$job['endtime'].'","schedtime":"'.$job['schedtime'].'"}';
+            $item = '{"x":"'.$job['name'].'","y":["'.$starttime.'","'.$endtime.'"],"fillColor":"'.$fillcolor.'","name":"'.$job['name'].'","jobid":"'.$job['jobid'].'","starttime":"'.$job['starttime'].'","endtime":"'.$job['endtime'].'","schedtime":"'.$job['schedtime'].'","client":"'.$job['client'].'"}';
             array_push($jobs, json_decode($item));
 
         }
@@ -744,7 +761,7 @@ class JobController extends AbstractActionController
         $result = $jobs;
 
       } catch(Exception $e) {
-        echo $e->getMessage();
+        echo $e->getMessages();
       }
     }